requirements.txt
fastapi
uvicorn
python-dotenv
langchain-core
langchain-community
langchain-milvus
langgraph
dashscope
pymilvus
redis
pypdf
sentence-transformers
torch
numpy
ragas
rank-bm25
transformers
sentence-transformers